Commit singing code.
This is the code from https://learn.adafruit.com/adabox002/make-your-robot-sing.

I've reinstated the license comment at the top (which was not on the
web page).  Otherwise, this commit is that page's code as-is.  It
looks like it diverged from the Github repo's code in a few minor
formatting aspects.