Reorder
Rearrange everything to make it easier to maintain.

The code was pretty scattered around.  This commit rearranges it.  The
only thing this commit does is to move existing code around
vertically; it doesn't change any of it.  It does add a few section
header comments.

The upstream version doesn't work with Arduino-Makefile; some
functions are used before they're declared.  It seems that there's
some magic that the Arduino IDE does that Arduino-Makefile doesn't.
(IIRC, the IDE wraps everything in one big class.)  This organization
makes sure that everything is defined (and hence declared) before it's
used.